refactor dao proxy
