+++ src/Dao/DaoProxyInterface.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Codeages\Biz\Framework\Dao;
+
+interface DaoProxyInterface
+{
+}
