+++ src/Dao/SerializerInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Codeages\Biz\Framework\Dao;
+
+interface SerializerInterface
+{
+    public function serialize($method, $value);
+
+    public function unserialize($method, $value);
+}
